Plot a single label with a central barcode and various fields of information to be printed on the label

Plot a single label with a central barcode and various fields of information to be printed on the label

Usage

make_single_label(
  layout,
  texts,
  labels,
  fontface,
  fontsize,
  bc = "",
  bctype = c("qr", "128"),
  inner.border = TRUE
)

Arguments

layout

a list of rectangle coordinates corresponding to boxes harbouring different label's parts

texts

a vector of fields text values information, same length as layout

labels

a vector of fields label, same length as layout

fontface

a vector of fontfaces, same length as layout

fontsize

a vector of fontsizes, same length as layout

bc

a single text value that will be used for the barcode

bctype

the type of barcode (either '128' or 'qc')

inner.border

should borders of individual rectangles be drawn (TRUE/FALSE)