Print labels as a pdf file

Print labels as a pdf file

Usage

print_labels(
  data,
  filename,
  layout,
  columns,
  bccol,
  bctype = "128",
  label_sizes,
  byrow = TRUE,
  fontface = "normal",
  fontsize = 6,
  inner.border = TRUE,
  outer.border = TRUE
)

Arguments

data

a data.frame with one row per label

filename

output filename

layout

the single label layout (see make_single_label function)

columns

the names of the columns in data to be used (same length as layout)

bccol

the name of the column to use for the barecode

bctype

the type of barcode (either '128' or 'qc')

label_sizes

the description of label format

byrow

TRUE or FALSE. In case of label pages should the page be filled by row or by column

fontface

font face (bold or plain)

fontsize

font size

inner.border

draw border inside label to separate fields

outer.border

draw rectangle around label