Single Trial Analysis report generation as Word document using officer

Single Trial Analysis report generation as Word document using officer

Usage

stareport(
  fit,
  file = file,
  template = templt,
  trialdesc = "",
  trialName = "",
  crop = "",
  outliers = NULL,
  excluded = NULL,
  spatial = FALSE,
  toc = TRUE,
  table.dec = 4
)

Arguments

fit

file

template

trialName

spatial

toc

table.dec